Corrigido a forma como as query strings são montadas nos services.
