+++ iOS-challenge-BTG/Services/MovieService.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class MovieService {
+class MovieService: Service {
 
     // MARK: - Requests
 
@@ -16,9 +16,13 @@ class MovieService {
         page: Int,
         completion: @escaping (MovieViewModel?, MovieServiceError?) -> ()) {
 
+        let query = [
+            URLQueryItem(name: "page", value: "\(page)")
+        ]
+
         let path = "/movie/popular"
 
-        guard let url = URL(string: "\(createApiUrl(with: path))&page=\(page)") else { return }
+        guard let url = createApiUrl(with: path, queryItems: query) else { return }
         URLSession.shared.dataTask(with: url, completionHandler: { (data, response, error) -> () in
             do {
                 if error != nil {
@@ -47,7 +51,7 @@ class MovieService {
 
         let path = "/movie/\(id)"
 
-        guard let url = URL(string: createApiUrl(with: path)) else { return }
+        guard let url = createApiUrl(with: path, queryItems: []) else { return }
         URLSession.shared.dataTask(with: url, completionHandler: { (data, response, error) -> () in
             do {
                 if error != nil {
@@ -103,23 +107,6 @@ class MovieService {
                 completion(nil, MovieServiceError.CannotFetch())
             }
         }).resume()
-    }
-
-    // MARK: - Utility
-
-    private func createApiUrl(with path: String) -> String {
-        return "\(Constants.api().url)\(path)?api_key=\(Constants.api().key)&language=\(Constants.api().language)"
-    }
-
-    private func createApiUrl(with path:String, queryItems: [URLQueryItem]) -> URL? {
-        var query = [
-            URLQueryItem(name: "api_key", value: Constants.api().key),
-            URLQueryItem(name: "language", value: Constants.api().language)
-        ]
-        query.append(contentsOf: queryItems)
-        var urlComponents = URLComponents(string: "\(Constants.api().url)\(path)")
-        urlComponents?.queryItems = query
-        return urlComponents?.url
     }
 }
